Add files via upload
In this phase, a server was developed to interact with the COM objects of the Cayman application using win32com.client. This server is then connected to a client using the Socket communication protocol. The aim is to collect data from the user interface of the Cayman and send it to the client. The client developed in this part is named "2eme_client.py".